修改缩略图集打包机制

+++ script/build.ts
@@ -39,7 +39,7 @@ import * as rollup from 'rollup';
                 });
             })
         );
-        await fse.remove('./dist/maps/');
+        if (process.argv[2] !== '1') await fse.remove('./dist/maps/');
         // 在线查看什么都看不到，这编辑器难道还需要留着吗？
         await fse.remove('./dist/_server');
         await fse.remove('./dist/editor.html');
